i#3995 multi-window: Convert delayed tracing to drbbdup
Converts the existing -trace_after_instrs delayed tracing feature to
use the drbbdup multi-instrumentation library with two cases: counting
instructions, and full tracing.

The drbbdup case encoding is a global std::atomic, written to using
language features which are lock-free and safe for client use.

This will lay the groundwork for the full i#3995 feature of repeatedly
swapping between the two cases.

TODO: For function tracing, we need to invoke drwrap only for the full
tracing case and not for the instruction counting case.  The plan is
to add a drwrap mode where drwrap does not use its own insertion event
and instead the user invokes drwrap from its insertion event.

TODO: For AArch64, drbbdup needs to handle reachability and encoding
issues with loading the global case value into a register.

Issue: #3995